+++ packages/frontend/src/components/SettingsValidator/ValidatorTable.tsx
@@ -1,5 +1,11 @@
-import { ValidationResponse } from '@lightdash/common';
-import { Alert, clsx, Flex, Table, Text, useMantineTheme } from '@mantine/core';
+import {
+    isChartValidationError,
+    isDashboardValidationError,
+    ValidationErrorChartResponse,
+    ValidationErrorDashboardResponse,
+    ValidationResponse,
+} from '@lightdash/common';
+import { Alert, Flex, Table, Text, useMantineTheme } from '@mantine/core';
 import {
     IconAlertCircle,
     IconLayoutDashboard,
@@ -17,25 +23,31 @@ const getLinkToResource = (
     validationError: ValidationResponse,
     projectUuid: string,
 ) => {
-    if (validationError.chartUuid)
+    if (isChartValidationError(validationError))
         return `/projects/${projectUuid}/saved/${validationError.chartUuid}`;
-    if (validationError.dashboardUuid) {
+
+    if (isDashboardValidationError(validationError))
         return `/projects/${projectUuid}/dashboards/${validationError.dashboardUuid}/view`;
-    }
 
     return;
 };
 
 const Icon = ({ validationError }: { validationError: ValidationResponse }) => {
-    if (validationError.chartUuid)
+    if (isChartValidationError(validationError))
         return getChartIcon(validationError.chartType);
-    if (validationError.dashboardUuid)
+    if (isDashboardValidationError(validationError))
         return <IconBox icon={IconLayoutDashboard} color="green.8" />;
     return <IconBox icon={IconTable} color="indigo.6" />;
 };
 
 const UpdatedAtAndBy: FC<
-    Required<Pick<ValidationResponse, 'lastUpdatedAt' | 'lastUpdatedBy'>>
+    Required<
+        | Pick<ValidationErrorChartResponse, 'lastUpdatedAt' | 'lastUpdatedBy'>
+        | Pick<
+              ValidationErrorDashboardResponse,
+              'lastUpdatedAt' | 'lastUpdatedBy'
+          >
+    >
 > = ({ lastUpdatedAt, lastUpdatedBy }) => {
     const timeAgo = useTimeAgo(lastUpdatedAt);
 
@@ -51,7 +63,7 @@ export const ValidatorTable: FC<{
     data: ValidationResponse[];
     projectUuid: string;
 }> = ({ data, projectUuid }) => {
-    const { classes } = useTableStyles();
+    const { cx, classes } = useTableStyles();
     const { colors } = useMantineTheme();
 
     const history = useHistory();
@@ -79,9 +91,8 @@ export const ValidatorTable: FC<{
 
     return (
         <Table
-            className={clsx(
+            className={cx(
                 classes.root,
-                classes.smallHeaderText,
                 classes.smallPadding,
                 classes.stickyHeader,
             )}
@@ -134,7 +145,11 @@ export const ValidatorTable: FC<{
                                   </Alert>
                               </td>
                               <td>
-                                  {validationError.lastUpdatedAt &&
+                                  {(isChartValidationError(validationError) ||
+                                      isDashboardValidationError(
+                                          validationError,
+                                      )) &&
+                                  validationError.lastUpdatedAt &&
                                   validationError.lastUpdatedBy ? (
                                       <UpdatedAtAndBy
                                           lastUpdatedAt={